chore(planx): move predicate lifting into a new submodule
