+++ planning/planx/src/lib.rs
@@ -5,12 +5,12 @@ pub mod errors;
 mod expressions;
 mod fluents;
 mod goals;
-pub mod lift_predicates;
 mod metrics;
 mod model;
 mod objects;
 mod params;
 pub mod pddl;
+pub mod preprocessing;
 mod sym;
 mod tasks;
 mod timing;
