+++ planning/planx/src/preprocessing/mod.rs
@@ -1,0 +1,3 @@
+mod lift_predicates;
+
+pub use lift_predicates::lift_predicates_to_state_functions;
